backend/pyproject.toml
[project]
name = "fleetflow-backend"
version = "1.0.0"
description = "FleetFlow: Modular Fleet & Logistics Management System — API Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.30.0",
    "alembic==1.14.1",
    "pydantic[email]==2.10.4",
    "pydantic-settings==2.7.1",
    "passlib[bcrypt]==1.7.4",
    "pyjwt==2.10.1",
    "python-multipart==0.0.20",
    "httpx==0.28.1",
    "slowapi==0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.0",
    "ruff==0.8.6",
    "mypy==1.14.1",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
